Refactored TestLauncher
