+++ src/main/java/com/testmaster/Main.java
@@ -1,14 +1,14 @@
 package com.testmaster;
 
-import com.testmaster.executor.TestExecutor;
+import com.testmaster.launcher.TestLauncher;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 public class Main {
   public static void main(String[] args) {
     ApplicationContext context = new ClassPathXmlApplicationContext("application-context.xml");
-    TestExecutor executor = context.getBean(TestExecutor.class);
+    TestLauncher launcher = context.getBean(TestLauncher.class);
 
-    executor.startApp();
+    launcher.startApp();
   }
 }
